Moving some files around and creating investment calculator components
