+++ app/ui/portfolio/modals/investmentCalculator/InvestmentInputRow.tsx
